finished tests for structured results on train epoch
