+++ tests/base/deterministic_model.py
@@ -103,43 +103,6 @@ class DeterministicModel(LightningModule):
         prototype_loss = outputs[0]
         return prototype_loss
 
-    # --------------------------
-    # Result returns
-    # --------------------------
-    def training_step_full_loop_result_obj(self, batch, batch_idx):
-        """
-        Full loop flow train step
-        """
-        x, y = batch
-        x = x.view(x.size(0), -1)
-        y_hat = self(x)
-        loss_val = self.loss(y, y_hat)
-        result = TrainResult(minimize=loss_val)
-        result.log('train_step_acc1', loss_val + 1)
-        self.training_step_called = True
-        return result
-
-    def training_step_end_full_loop_result_obj_dp(self, result):
-        """
-        Full loop flow train step
-        """
-        self.assert_backward = False
-
-        result.minimize = result.minimize.mean()
-        result.checkpoint_on = result.checkpoint_on.mean()
-        result.train_step_acc1 = result.train_step_acc1.mean()
-        result.log('train_step_end_acc1', 1)
-        self.training_step_end_called = True
-        return result
-
-    def training_epoch_end_full_loop_result_obj(self, result):
-        """
-        Full loop flow train step
-        """
-        result.log('train_epoch_end_acc1', 1)
-        self.training_epoch_end_called = True
-        return result
-
     def training_step_no_default_callbacks_for_train_loop(self, batch, batch_idx):
         """
         Early stop and checkpoint only on these values
