+++ tests/base/model_template.py
@@ -63,6 +63,9 @@ class EvalModelTemplate(
         self.hidden_dim = hidden_dim
         self.b1 = b1
         self.b2 = b2
+        self.training_step_called = False
+        self.training_step_end_called = False
+        self.training_epoch_end_called = False
 
         # if you specify an example input, the summary will show input/output for each layer
         # TODO: to be fixed in #1773
